requirements.txt
transformers
sentence-transformers
faiss-cpu
PyMuPDF
fastapi
uvicorn
python-multipart
